Add new recordingStatus property
Add a new property to the AudioRecorder object for returning the current
recording status. Currently this includes whether the recorder is
recording, and if so, the current duration of the recording.

The duration is more accurate on iOS than on Android, because on Android
we manually take a start timestamp, which may not coincide precisely
with the time that the native recorder starts recording.